% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCdisslist.R
\name{MCdisslist}
\alias{MCdisslist}
\title{List of dissimilarity matrices}
\usage{
MCdisslist(
  MCrseqdata,
  method = "LCS",
  seqref = NULL,
  full.matrix = FALSE,
  use.udiss = FALSE,
  ...
)
}
\arguments{
\item{MCrseqdata}{List of state sequence objects of class \code{stslist}.}

\item{method}{string. Name of a distance method (see \code{\link[TraMineR]{seqdist}}).}

\item{seqref}{state sequence object of class \code{stslist}. Fixed reference sequences.}

\item{full.matrix}{logical. Should pairwise distances be returned in matrix form? If \code{FALSE} (default),
a list of \code{dist} objects is returned. Applies only when \code{seqref=NULL}.}

\item{use.udiss}{logical. Should computation be based on unique sequences?}

\item{...}{further arguments passed to \code{seqdist}.}
}
\value{
list of dissimilarity matrices or \code{dist} objects with logical attribute \code{"obs"}, which is \code{TRUE} when the list includes the dissimilarities between observed sequences as last element.
}
\description{
Compute the dissimilarity matrix for each of the provided sets of sequences.
}
\details{
When \code{use.udiss=TRUE}, the function first computes dissimilarities between unique merged replicated sequences through a single call to \code{seqdist()} and the set of dissimilarity matrices are then extracted from the resulting distance matrix. This is generally faster when the number of unique merged replicated sequences is less than \code{sqrt(number of replicated datasets) * (sample size)}, which can be checked with \code{\link{MCnunique}}.
}
\seealso{
\code{\link{MCseqReplicate}}, \code{\link{MCudist}} and examples in their help pages.
}
