% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCudist.R
\name{MCExtractDist}
\alias{MCExtractDist}
\title{Extract k-th dissimilarity matrix from u.diss}
\usage{
MCExtractDist(u.diss, k, full.matrix = FALSE)
}
\arguments{
\item{u.diss}{\code{u.diss} object returned by \code{MCudist}: dissimilarities between unique replicated sequences.}

\item{k}{integer. Subset index number for which the dissimilarity matrix must be extracted}

\item{full.matrix}{logical. If \code{FALSE}, the distance matrix is returned as a \code{dist} object. Ignored for distances to reference sequences.}
}
\value{
a dissimilarity matrix or distance object.
}
\description{
Extract k-th dissimilarity matrix from u.diss
}
\seealso{
\code{\link{MCudist}}
}
