% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCudist.R
\name{MCnunique}
\alias{MCnunique}
\title{Number of unique replicated sequences}
\usage{
MCnunique(MCrseqdata, check = FALSE)
}
\arguments{
\item{MCrseqdata}{list of replicated \code{stslist} state sequence datasets (all of same size and with same alphabet.}

\item{check}{logical. When \code{TRUE}, check if the number of unique replicated sequences is less than  \code{sqrt(number of replicated datasets) * (sample size)}?}
}
\value{
\code{nu} number of unique replicated sequences and, when \code{check=TRUE}, \code{u.ok} the check result.
}
\description{
Number of unique replicated sequences
}
\seealso{
\code{\link{MCudist}}, \code{\link{MCseqdistSE}}
}
