% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCratios.R
\name{MCratios}
\alias{MCratios}
\title{Ratios of distances on their standard errors}
\usage{
MCratios(object, diss.o = NULL)
}
\arguments{
\item{object}{Object of class \code{distMC} as generated by \code{\link{MCseqdistSE}} and \code{\link{seqdistMCSE}}.}

\item{diss.o}{Matrix or \code{dist} object. Pairwise dissimilarities between observed sequences.}
}
\value{
\code{diss.z}, \code{MC.mean.z}, and \code{mean.se} (the three as \code{dist} objects).
}
\description{
Ratios of the observed distances to their MC standard errors and of the mean MC-simulated distances to the standard error of the mean.
}
\details{
The standard error of the mean simulated distances is \code{mean.se = MC.se/sqrt(R)} (or \code{mean.se = MC.sd/R} when object is obtained with \code{seqdistMCSE}, because there are \code{R*R} simulated distances in that case). The ratios computed are \code{diss.z = diss.o/MC.se}, where \code{diss.o} is the distance between observed sequences, and \code{MC.mean.z = MC.mean/mean.se} with \code{MC.mean} the mean of the MC-simulated distances.

When \code{diss.o=NULL}, the \code{diss.o} element of \code{object} is used when it exists.

This function is handy to get afterwards ratios for outcome of \code{\link{seqdistMCSE}} obtained with \code{ratios=FALSE}.
}
\seealso{
\code{\link{MCseqdistSE}} and \code{\link{print.MCratios}}.
}
\author{
Gilbert Ritschard
}
