% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCudist.R
\name{MCudist}
\alias{MCudist}
\title{Dissimilarities between unique replicated sequences}
\usage{
MCudist(MCrseqdata, method = "LCS", seqref = NULL, ...)
}
\arguments{
\item{MCrseqdata}{list of replicated \code{stslist} state sequence datasets (all of same size and with same alphabet)}

\item{method}{string. Name of distance method (see \code{seqdist}).}

\item{seqref}{state sequence object of class \code{stslist}. Fixed reference sequences.}

\item{...}{Further arguments passed to \code{seqdist}}
}
\value{
object of class \code{u.diss} (pairwise dissimilarities between unique sequences) with two attributes: \code{sdx}, inverted aggregation indexes, \code{N}, number of datasets, and \code{obs}, logical indicating whether \code{k=N} corresponds to observed sequences.
}
\description{
Returns the dissimilarity matrix (or \code{dist} object) between merged replicated sequences with the disaggregation indexes as attribute.
}
\examples{
## mini test data, 6 sequences of length 4, 4 unique sequences
exdata <- read.table(text="
                a a b b
                a a b b
                b b a a
                a c c b
                b b a c
                b b a c
                ")
weights=rep(1, nrow(exdata))
s.exdata <- seqdef(exdata, weights = weights, id=paste("id",1:nrow(exdata), sep=""))

## 3 altered sequence datasets
(altseq.list <- MCseqReplicate(s.exdata, J=1, R=3))

MCnunique(altseq.list, check=TRUE)


u.diss <- MCudist(altseq.list, method="LCS", full.matrix=FALSE)
## Dissimilarities within first MC-set
MCExtractDist(u.diss, 1)

## list of dissimilarity matrices
disslist <- MCdisslist(altseq.list, use.udiss=TRUE)

}
\seealso{
\code{\link{MCExtractDist}}
}
