% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCmethods.R
\name{print.distMC}
\alias{print.distMC}
\title{Print method for distMC objects}
\usage{
\method{print}{distMC}(x, n = 6, what = "all", ...)
}
\arguments{
\item{x}{\code{distMC} object as returned by \code{MCseqdistSE}.}

\item{n}{Integer. Number of first sequences. Default is 6. If \code{n==0} or there are less than \code{n} sequences, results are printed for all pairs of sequences.}

\item{what}{character string. One of \code{"mean"}, \code{"sd"}, \code{"bias"}, \code{"both"}, and \code{"all"} (default). When \code{"all"}, ratios, when present are printed together with the mean and standard deviation. When \code{"both"}, means and standard deviations are printed.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Last printed table, a \code{matrix} when \code{toref} attribute is \code{TRUE} and a \code{dist} object otherwise.
}
\description{
Prints, for each pair of the first \code{n} sequences, the mean and/or the standard deviation of the MC-replicated distances between sequences. When available, ratios are also printed by default.
}
\seealso{
\code{\link{MCseqdistSE}}, \code{\link{summary.distMC}}.
}
\author{
Gilbert Ritschard
}
