% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCmethods.R
\name{print.MCratios}
\alias{print.MCratios}
\title{Print method for MCratios objects}
\usage{
\method{print}{MCratios}(x, n = 6, what = "all", ...)
}
\arguments{
\item{x}{\code{MCratios} object as returned by \code{MCratios}.}

\item{n}{Integer. Number of first sequences. Default is 6. If \code{n==0} or there are less than \code{n} sequences, results are printed for all pairs of sequences.}

\item{what}{character string. One of \code{"all"} (default), \code{"diss"}, \code{"mean"}, and \code{"se"} .}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Last printed table, a \code{matrix} when \code{toref} attribute is \code{TRUE} and a \code{dist} object otherwise.
}
\description{
Prints ratios for each pair of the first \code{n} sequences.
}
\seealso{
\code{\link{MCratios}}. %\code{\link{seqdistMCSE}},
}
\author{
Gilbert Ritschard
}
