% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCmethods.R
\name{summary.distMC}
\alias{summary.distMC}
\title{Summary method for distMC objects}
\usage{
\method{summary}{distMC}(object, ..., silent = FALSE)
}
\arguments{
\item{object}{\code{distMC} object as returned by \code{MCseqdistSE}.}

\item{...}{further arguments passed to or from other methods.}

\item{silent}{logical: Should additional info be displayed?}
}
\value{
\code{fivenumb} table with the statistics (min, Q1, med, Q3, max) of the observed dissimilarities, the mean, standard deviation, and standard error of the MC-simulated dissimilarities, standardized ratios, MC-bias and mean squared errors when available.
}
\description{
Prints summary statistics of the observed dissimilarity \code{diss}, the mean \code{MC.mean}, standard deviation \code{MC.sd}, and standard error of dissimilarities between MC-replicated sequences, and the ratios  \code{diss/MC.se} and \code{MC.mean/MC.se}. Reported statistics concern all distances between original sequences.
}
\seealso{
\code{\link{MCseqdistSE}}, \code{\link{print.distMC}}
}
\author{
Gilbert Ritschard
}
