% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCmethods.R
\name{summary.MCratios}
\alias{summary.MCratios}
\title{Summary method for MCratios objects}
\usage{
\method{summary}{MCratios}(object, ..., weights = NULL, silent = FALSE, thresh = 2)
}
\arguments{
\item{object}{\code{MCratios} object as returned by \code{MCratios}.}

\item{...}{further arguments passed to or from other methods.}

\item{weights}{vector of doubles. Case weights.}

\item{silent}{logical: Should additional info be displayed?}

\item{thresh}{real: threshold for counting ratios less than \code{thresh}}
}
\value{
\code{fivenumb} table with the statistics (min, Q1, med, Q3, max) of \code{mean.se} and the standardized ratios \code{diss.z} and \code{MC.mean.z}.
}
\description{
Prints summary statistics of the ratios  \code{diss/MC.se} and \code{MC.mean/MC.se}. Reported statistics concern all distances between original sequences.
}
\seealso{
\code{\link{MCseqdistSE}}, \code{\link{print.distMC}}
}
\author{
Gilbert Ritschard
}
